Rename runGraph to Graph.
Cleanup branch crud
